+++ src/commands/interop/installer-module.ts
@@ -1,8 +1,9 @@
 import { Commander } from "../../functions/cli.ts";
 import { CheckForPath, JoinPaths, ParsePath } from "../../functions/filesystem.ts";
+import type { MANAGER_JS } from "../../types/platform.ts";
 
 export const Installers = {
-    UniJs: async (path: string, manager: "bun" | "npm" | "pnpm" | "yarn" | "deno") => {
+    UniJs: async (path: string, manager: MANAGER_JS) => {
         Deno.chdir(ParsePath(path));
         await Commander(manager, ["install"]);
         return;