+++ src/main.ts
@@ -109,7 +109,7 @@ const flags = Deno.args.map((arg) =>
 
 export const __FKNODE_SHALL_WE_DEBUG = flags.some((s) => StringUtils.normalize(s) === "fkndbg");
 DEBUG_LOG("Initialized __FKNODE_SHALL_WE_DEBUG constant (ENTRY POINT)");
-DEBUG_LOG("ARGS", flags)
+DEBUG_LOG("ARGS", flags);
 
 if (!StringUtils.validate(flags[0])) {
     try {